remove EigenVecWrapper (a duplicate of EigenVector)
